# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  version: 1.0.0
  description: >
    REST API para el sistema de gestión de pedidos, reseñas y analítica de
    restaurantes sobre MongoDB Atlas (CC3089 Bases de Datos 2).


    **Recursos principales**: Users, Restaurants, Menu Items, Carts, Orders,
    Reviews, Files (GridFS), Analytics.
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /restaurants:
    post:
      summary: Crear restaurante
      tags: [Restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, location, address, operatingHours]
              properties:
                name:
                  type: string
                  example: La Pizzería Artesanal
                description:
                  type: string
                location:
                  $ref: '#/components/schemas/GeoPoint'
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    zone:
                      type: string
                operatingHours:
                  type: object
                  description: Horarios por día de la semana.
                cuisineTypes:
                  type: array
                  items:
                    type: string
                  example: [italiana, mexicana]
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Restaurante creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'

    get:
      summary: Listar restaurantes
      tags: [Restaurants]
      parameters:
        - in: query
          name: cuisine
          schema:
            type: string
          description: Filtrar por tipo de cocina.
        - in: query
          name: isActive
          schema:
            type: string
            enum: ['true', 'false']
        - in: query
          name: isAcceptingOrders
          schema:
            type: string
            enum: ['true', 'false']
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista paginada de restaurantes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
  /restaurants/search:
    get:
      summary: Buscar restaurantes por ubicación (geoespacial)
      description: >
        Usa $geoIntersects sobre delivery_zones para encontrar restaurantes que
        cubren la ubicación del usuario, enriquecidos con restaurant_stats.
      tags: [Restaurants]
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
          description: Latitud del usuario.
        - in: query
          name: lng
          required: true
          schema:
            type: number
          description: Longitud del usuario.
        - in: query
          name: cuisine
          schema:
            type: string
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Restaurantes con zona de entrega que cubre la ubicación.
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  total:
                    type: integer
        '400':
          description: lat y lng son requeridos.
  /restaurants/many:
    post:
      summary: Crear varios restaurantes (insertMany)
      tags: [Restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurants]
              properties:
                restaurants:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      location:
                        $ref: '#/components/schemas/GeoPoint'
                      address:
                        type: object
                      operatingHours:
                        type: object
                      cuisineTypes:
                        type: array
                        items:
                          type: string
      responses:
        '201':
          description: Restaurantes creados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  insertedCount:
                    type: integer
  /restaurants/{id}:
    get:
      summary: Obtener restaurante por ID (con stats)
      tags: [Restaurants]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Restaurante con estadísticas pre-computadas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Restaurante no encontrado.

    patch:
      summary: Actualizar restaurante
      tags: [Restaurants]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                operatingHours:
                  type: object
                cuisineTypes:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Restaurante actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '404':
          description: Restaurante no encontrado.

    delete:
      summary: Eliminar restaurante
      tags: [Restaurants]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Restaurante eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '404':
          description: Restaurante no encontrado.
  /restaurants/{id}/status:
    patch:
      summary: Cambiar estado del restaurante (isAcceptingOrders / isActive)
      tags: [Restaurants]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isAcceptingOrders:
                  type: boolean
                isActive:
                  type: boolean
      responses:
        '200':
          description: Estado actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
  /restaurants/{id}/menu-categories:
    get:
      summary: Categorías de menú del restaurante (distinct)
      tags: [Restaurants]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Array de categorías únicas.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: [Entradas, Platos Principales, Postres, Bebidas]
...
  /orders:
    post:
      summary: Crear pedido (transacción multi-documento)
      description: >
        Valida carrito → ítems disponibles → restaurante abierto → zona de entrega
        ($geoIntersects) → calcula totales → inserta orden → elimina carrito →
        actualiza orderHistory del usuario. Todo dentro de una sesión transaccional.
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, cartId, deliveryAddress, paymentMethod]
              properties:
                userId:
                  type: string
                cartId:
                  type: string
                deliveryAddress:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    zone:
                      type: string
                    coordinates:
                      $ref: '#/components/schemas/GeoPoint'
                paymentMethod:
                  type: string
                  enum: [card, cash, transfer]
      responses:
        '201':
          description: Orden creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Error de validación (carrito vacío, restaurante cerrado, fuera de zona, etc.).

    get:
      summary: Listar órdenes ($lookup users + restaurants)
      tags: [Orders]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready_for_pickup, picked_up, delivered, cancelled]
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista de órdenes con datos enriquecidos de usuario y restaurante.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/cancelled:
    delete:
      summary: Eliminar órdenes canceladas (deleteMany)
      tags: [Orders]
      parameters:
        - in: query
          name: before
          schema:
            type: string
            format: date-time
          description: Eliminar canceladas creadas antes de esta fecha.
      responses:
        '200':
          description: Órdenes canceladas eliminadas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
  /orders/{id}:
    get:
      summary: Obtener orden por ID
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Detalle de la orden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Orden no encontrada.

    delete:
      summary: Eliminar orden
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Orden eliminada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '404':
          description: Orden no encontrada.
  /orders/{id}/status:
    patch:
      summary: Transición de estado con statusHistory
      description: >
        Máquina de estados: pending → confirmed → preparing → ready_for_pickup →
        picked_up → delivered. Cualquier estado intermedio puede → cancelled.
        Registra cada transición en statusHistory embebido con timestamp y duración.
      tags: [Orders]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [confirmed, preparing, ready_for_pickup, picked_up, delivered, cancelled]
                actor:
                  type: string
                  default: system
                reason:
                  type: string
                  description: Razón de cancelación (solo si status = cancelled).
      responses:
        '200':
          description: Transición exitosa.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  transition:
                    type: string
                    example: "pending → confirmed"
        '400':
          description: Transición no permitida.
        '404':
          description: Orden no encontrada.
...
components:
  parameters:
    objectId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: ObjectId de MongoDB (24 caracteres hex).
    skip:
      in: query
      name: skip
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    GeoPoint:
      type: object
      properties:
        type:
          type: string
          example: Point
        coordinates:
          type: array
          items:
            type: number
          example: [-90.5069, 14.5943]
          description: "[longitude, latitude]"
...
    Restaurant:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        location:
          $ref: '#/components/schemas/GeoPoint'
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            zone:
              type: string
        operatingHours:
          type: object
        cuisineTypes:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isAcceptingOrders:
          type: boolean
        logoFileId:
          type: string
          nullable: true
        menuItemCount:
          type: integer
...
    Order:
      type: object
      properties:
        _id:
          type: string
        orderNumber:
          type: string
          example: ORD-1710000000000-123
        userId:
          type: string
        restaurantId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
              name:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
              subtotal:
                type: number
        deliveryAddress:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            zone:
              type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, ready_for_pickup, picked_up, delivered, cancelled]
        statusHistory:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              timestamp:
                type: string
                format: date-time
              actor:
                type: string
              durationFromPrevSec:
                type: integer
        subtotal:
          type: number
        tax:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [card, cash, transfer]
        cancellationReason:
          type: string
          nullable: true
        estimatedDelivery:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateResult:
      type: object
      properties:
        updated:
          type: integer
          example: 1
    DeleteResult:
      type: object
      properties:
        deleted:
          type: integer
          example: 1
...